requirements.txt
# this is the "requirements.txt" file...

python-dotenv

requests # used to parse data from Spotify API

flask # used for web app
gunicorn # used for HEROKU DEPLOY; Heroku server wants to use this package to run our web app

# using this to find and run our tests:
pytest

# using bsoup in testing the web app:
beautifulsoup4
